% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSetupRun.R
\name{defineParameters}
\alias{defineParameters}
\title{Define CWatM settings & parameters}
\usage{
defineParameters(pathRoot, overwrite = FALSE)
}
\arguments{
\item{pathRoot}{a character vector with the path to the CWatM root folder}

\item{overwrite}{logical. Should the function  overwrite existing files?}
}
\value{
an MS Excel settings spreadsheet (see Details)
}
\description{
This function creates a default settings spreadsheet, granting the user with complete control over all model parameters.
}
\details{
The \code{pathRoot} variable indicates the root folder that stores CWatM files.
The output settings file path is: \code{pathRoot/cwatm_settings/cwatm_settings.xlsx}, and the user should make sure that path exists.
}
\examples{
\dontrun{
defineParameters(pathRoot = "C:/IIASA/cwatm", overwrite = TRUE)
}
}
