% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{getSettingsDomains}
\alias{getSettingsDomains}
\title{Get all domains names from the default settings file}
\usage{
getSettingsDomains()
}
\value{
A character vector with all domains of the CWatM settings file.
}
\description{
This function provides a charachter vector with all domains (i.e., sections) of the CWatM settings file.
}
\examples{
getSettingsDomains()

}
