% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{getSettingsTable}
\alias{getSettingsTable}
\title{Get all domains names from the default settings file}
\usage{
getSettingsTable(settingsfile = NULL, domain)
}
\arguments{
\item{settingsfile}{A character vector with a path to an MS Excel settings file or to a \code{settings.ini} file. Defaults to \code{NULL}}

\item{domain}{A character vector with a settings file domain name}
}
\value{
A \code{data.frame}  with all variables, values, and comments for a given domain (see Details).
}
\description{
This function provides a table with  all variables and values for a given domain of the CWatM settings file.
}
\details{
The function extracts data from a user-defined settings file. In case a user does not define a settings file (\code{settingsfile = NULL}),
the results show a default settings file.
}
\examples{
getSettingsTable(settingsfile = NULL, domain = "FILE_PATHS")
dontrun {
getSettingsTable(settingsfile = "C:/IIASA/cwatm/cwatm_settings/cwatm_settings.xlsx", domain = "OPTIONS")
getSettingsTable(settingsfile = "C:/IIASA/cwatm/cwatm_settings/cwatm_settings_sorekBasin.ini", domain = "OPTIONS")
}

}
