% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{raster_as_nc}
\alias{raster_as_nc}
\title{Converts raster to netCDF}
\usage{
raster_as_nc(
  raster_in,
  time,
  nc_attrs,
  pathfile_out = "../output.nc",
  loud = FALSE,
  fill.value = NaN
)
}
\arguments{
\item{raster_in}{an input \code{RasterLayer}, \code{RasterStack} or a \code{list}}

\item{time}{a vector of class \code{Date} for the temporal dimension of the dataset (Default: NULL; see Details)}

\item{nc_attrs}{a named list with the variable and global attributes to be written to the NetCDF file (see Details)}

\item{pathfile_out}{a path and a name for the output NetCDF to be written}

\item{loud}{\code{logical.} Do you want to print the \code{ncdf4} object to consule?}

\item{fill.value}{NoData value (Default: NaN)}
}
\value{
an NetCDF file (see Details)
}
\description{
This function converts a \code{RasterLayer}, \code{RasterStack} or a \code{list} of these classes to a 'NetCDF' file.
}
\details{
This function accepts spatio-temporal data and export it to a NetCDF file.
For spatio-temporal data, a list of \code{RasterLayer}/\code{RasterStack} objects is expected. Each list item stands for a point in time, and
each \code{RasterLayer} stands for the spatial dimensions and data of a variable. If multiple variable are required, a \code{RasterStack} can be used, so
each layer stands for a variable.
In case  temporal dimension is required, the user can use the 'time' argument to specify the time points for the data. This argument
accepts a vector of the class \code{Date} and assumes the origin is 01-01-1901.
The 'nc_attrs' argument accepts a named list with the following variable attributes: 'dname', 'dlname', 'units'
and 'prec', and some of the global attributes: 'history', 'source_software', 'title', 'keywords', 'source', and 'institution'.
The variable name (dname) and long name (dlname) expects a character vector of the same length of the number of variables, or \code{NULL}.
If \code{NULL} is assigned, the variable name is taken from the \code{RasterLayer} name.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
