% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{settings_ini2list}
\alias{settings_ini2list}
\title{Import a cwatm .ini setting files to a settings List}
\usage{
settings_ini2list(inipath = "./settings_cwatm.ini")
}
\arguments{
\item{inipath}{a path to the input \code{.ini} settings file.}
}
\value{
\code{list}. A names setting \code{list} )
}
\description{
This function imports a cwatm .ini settings file and save it as a named settings \code{list}
}
\details{
This function reads a cwatm .ini file from a path and saves it as a named list. Each list item stands for a settings domain, in which
all variable definitions are ketp in a \code{data.frame} (with 3 columns: variable, value, comment). The funtion does not import comment though.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
