% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSetupRun.R
\name{settings_searchFiles}
\alias{settings_searchFiles}
\title{Function to serach and automatically upadte values in a CWatM named settings list}
\usage{
settings_searchFiles(settingsList = NULL, ...)
}
\arguments{
\item{settingsList}{named \code{list} with CWatM settings}

\item{...}{additional arguments (see details)}
}
\value{
named \code{list}; a CWatM settings list.
}
\description{
This function search user-defined folders for CWatM data inputs and automatically updates a settings named \code{list}
}
\details{
This function assumes that the CWatM data folders' structure and file names follows the convention described in the models' documentation:
\url{https://cwatm.iiasa.ac.at/}. The user can provoke a call to  \code{\link[=settings_setFilePaths]{settings_setFilePaths()}}, if \code{character} FILE_PATHS are provided.
In order to search for a specific file, the function uses a set of pre-defined variable-specific search expressions.
When updated with a new file path; the followin comment ("Updated") is appended to the settings table. If the search results with multiple matchs, the first one is updated
and the following comment is appended to the settings table: "Multiple results were found; taken the first one".
}
\examples{
\dontrun{
my_settings <- settings_ini2list(inipath = "./cwatm/settings_template.ini")

# set user-defined paths
root = "d:/cwatm/model_data"
maps = "d:/cwatm/model_data/maps"
meteo = "./climate"

my_settings <- settings_searchFiles(settingsList = my_settings,
                                     PathRoot = root,
                                     PathMaps = maps,
                                     PathMeteo = meteo)
}
}
