% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSetupRun.R
\name{settings_setFilePaths}
\alias{settings_setFilePaths}
\title{Update FILE_PATHS in a CWatM settings \code{list}}
\usage{
settings_setFilePaths(
  settingsList,
  PathRoot = NULL,
  PathOut = NULL,
  PathMaps = NULL,
  PathMeteo = NULL
)
}
\arguments{
\item{settingsList}{named \code{list} with CWatM settings}

\item{PathRoot}{\code{character}; path to folder containing all CWatM datasets}

\item{PathOut}{\code{character}; path to folder to save CWatM simulation's outputs}

\item{PathMaps}{\code{character}; path to folder containing all input data (excluding meteorological varaibales)}

\item{PathMeteo}{\code{character}; path to folder containing all meteorological variables}
}
\value{
named \code{list}; a CWatM settings list.
}
\description{
/
This function updates the FILE_PATHS domain of a CWatM settings named \code{list} based on user-defined values.
}
\details{
The function updates the settings file FILE_PATHS domain allowing the user to define its own paths to the model data and outputs.
By convention, both \code{PathMaps} and \code{PathMeteo} are sub-folders of \code{PathRoot}. Users are encourages to provide each of these argument with the
full pat. if a sub-folder of \code{PathRoot} the function rewrites the path as: "$(PathRoot)/Path_to_data". Alternatively, the user can use the following format to
provide a relative path: "./Path_to_data", whereas the "./" stands for the current working directory.
}
\examples{
\dontrun{

# set workinf directory
setwd("d:/cwatm/model_data")

# load a settings file template to a named `list`
my_settings <- settings_ini2list(inipath = "./cwatm/settings_template.ini")

# set user-defined paths
root = "d:/cwatm/model_data"
maps = "d:/cwatm/model_data/maps"
meteo = "./climate"

my_settings <- settings_setFilePaths(settingsList = my_settings,
                                     PathRoot = root,
                                     PathOut = NULL,
                                     PathMaps = maps,
                                     PathMeteo = meteo)
}/
}
