% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{settings_list2spreadsheet}
\alias{settings_list2spreadsheet}
\title{Define CWatM settings & parameters}
\usage{
settings_list2spreadsheet(
  settingsList = NULL,
  output_path = "./cwatm_settings.xlsx",
  overwrite = FALSE
)
}
\arguments{
\item{settingsList}{named list, a settingsList object (Defaults to \code{NULL}; see Details)}

\item{output_path}{character vector, the settings spreadsheet's path and name}

\item{overwrite}{logical. Should the function  overwrite existing files?}
}
\value{
an MS Excel settings spreadsheet (see Details)
}
\description{
This function creates a default settings spreadsheet, granting the user with complete control over all model parameters.
}
\details{
The user can choose a settingList object (e.g., as in the output of \code{settings_ini2list()}). Otherwise (if \code{NULL}) the function uses a default settingsList
}
\examples{
\dontrun{
settings_list2spreadsheet( output_path = "./cwatm_settings.xlsx", overwrite = TRUE)
}
}
